Add files via upload
This is a code for a two-person game. First, Player 1 will input three movie names. The computer will then randomly select one of the movies and replace random letters in the movie name with a placeholder (for example, * or #). This new mystery word will be printed on the screen, and Player 2 will then have to guess the actual movie name!